DB TO Excel

+++ src/main/java/com/abdo/springbatchcustomer/repo/RevenueRepo.java
@@ -3,5 +3,5 @@ package com.abdo.springbatchcustomer.repo;
 import com.abdo.springbatchcustomer.entity.Revenue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RevenueRepo extends JpaRepository<Revenue, Integer> {
+public interface RevenueRepo extends JpaRepository<Revenue, Long> {
 }
